after add skeleton React and add React context

+++ src/App.js
@@ -5,31 +5,37 @@ import Header from './components/Header';
 import Drawer from './components/Drawer';
 import Home from './components/pages/Home';
 import Favorites from './components/pages/Favorites';
+import AppContext from './context';
 
 function App() {
   const [items, setItems] = React.useState([]);
   const [cartItems, setCartItems] = React.useState([]);
   const [favorites, setFavorites] = React.useState([]);
   const [cartOpened, setCartOpened] = React.useState(false);
+  const [isLoading, setIsLoading] = React.useState(true);
   const [searchValue, setSearchValue] = React.useState('');
 
   React.useEffect(() => {
     async function fetchData() {
-      await axios
-        .get('https://63df9565a76cfd410584dbbf.mockapi.io/catr')
-        .then(res => setCartItems(res.data));
-      await axios
-        .get('https://63df9565a76cfd410584dbbf.mockapi.io/items')
-        .then(res => setItems(res.data));
+      const itemsRes = await axios.get(
+        'https://63df9565a76cfd410584dbbf.mockapi.io/items'
+      );
+      const cartRes = await axios.get(
+        'https://63df9565a76cfd410584dbbf.mockapi.io/cart'
+      );
+
+      setCartItems(cartRes.data);
+      setItems(itemsRes.data);
+      setIsLoading(false);
     }
     fetchData();
   }, []);
 
   const onAddToCart = async obj => {
-    if (cartItems.find(item => item.title === obj.title)) return;
+    if (!!cartItems.find(item => item.title === obj.title)) return;
     try {
       const res = await axios.post(
-        'https://63df9565a76cfd410584dbbf.mockapi.io/catr',
+        'https://63df9565a76cfd410584dbbf.mockapi.io/cart',
         obj
       );
       setCartItems(prev => [...prev, res.data]);
@@ -39,9 +45,10 @@ function App() {
   };
 
   const onRemoveToCart = async obj => {
+    if (!!!cartItems.find(item => item.title === obj.title)) return;
     try {
       await axios.delete(
-        `https://63df9565a76cfd410584dbbf.mockapi.io/catr/${obj.id}`
+        `https://63df9565a76cfd410584dbbf.mockapi.io/cart/${obj.id}`
       );
       setCartItems(prevState => {
         return [...prevState.filter(value => value?.id !== obj.id)];
@@ -63,40 +70,55 @@ function App() {
       setFavorites(prevState => [...prevState, obj]);
     }
   };
+
+  const isItemAdded = title => {
+    return cartItems.some(item => item.title === title);
+  };
   return (
-    <div className="wrapper clear">
-      {cartOpened && (
-        <Drawer
-          cardItems={cartItems}
-          onClickClose={() => setCartOpened(false)}
-          onClickRemove={obj => onRemoveToCart(obj)}
-        />
-      )}
-      <Header onClickCart={() => setCartOpened(true)} />
-      <Routes>
-        <Route
-          path="/"
-          element={
-            <Home
-              items={items}
-              cartItems={cartItems}
-              searchValue={searchValue}
-              setSearchValue={setSearchValue}
-              onChangeSearchInput={onChangeSearchInput}
-              onAddToCart={onAddToCart}
-              onRemoveToCart={onRemoveToCart}
-              onAddToFavorite={onAddToFavorite}
-            />
-          }
-        />
-        <Route
-          path="/favorites"
-          element={
-            <Favorites items={favorites} onAddToFavorite={onAddToFavorite} />
-          }
-        />
-      </Routes>
-    </div>
+    <AppContext.Provider
+      value={{
+        items,
+        cartItems,
+        favorites,
+        setCartOpened,
+        setCartItems,
+        isItemAdded,
+        onAddToCart,
+        onRemoveToCart,
+        onAddToFavorite,
+      }}
+    >
+      <div className="wrapper clear">
+        {cartOpened && (
+          <Drawer
+            cardItems={cartItems}
+            onClickClose={() => setCartOpened(false)}
+            onClickRemove={obj => onRemoveToCart(obj)}
+          />
+        )}
+        <Header onClickCart={() => setCartOpened(true)} />
+        <Routes>
+          <Route
+            path="/"
+            element={
+              <Home
+                searchValue={searchValue}
+                setSearchValue={setSearchValue}
+                isLoading={isLoading}
+                onChangeSearchInput={onChangeSearchInput}
+                onAddToCart={onAddToCart}
+                onRemoveToCart={onRemoveToCart}
+                onAddToFavorite={onAddToFavorite}
+              />
+            }
+          />
+          <Route
+            path="/favorites"
+            element={<Favorites onAddToFavorite={onAddToFavorite} />}
+          />
+        </Routes>
+      </div>
+    </AppContext.Provider>
   );
 }
